dl_gst_serve/requirements.txt
aniso8601==8.0.0
azure-core==1.4.0
azure-storage-blob==12.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
idna==2.9
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
msrest==0.6.13
numpy==1.18.2
oauthlib==3.1.0
pandas==1.0.3
pycparser==2.20
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
